+++ src/main/drivers/accgyro/accgyro_spi_bmi270.c
@@ -393,7 +393,7 @@ static bool bmi270GyroReadRegister(gyroDev_t *gyro)
         if (gyro->detectedEXTI > GYRO_EXTI_DETECT_THRESHOLD) {
 #ifdef USE_DMA
             if (spiUseDMA(dev)) {
-                dev->callbackArg = (uint32_t)gyro;
+                dev->callbackArg = (uintptr_t)gyro;
                 dev->txBuf[0] = BMI270_REG_ACC_DATA_X_LSB | 0x80;
                 gyro->segments[0].len = 14;
                 gyro->segments[0].callback = bmi270Intcallback;
